requirements.txt

Flask
paho-mqtt
tensorflow
ultralytics
scikit-learn
opencv-python
numpy
pandas
